FlaskServer/requirements.txt
flask
flask-socketio
flask-cors
gevent
psycopg2-binary